flasktaskr-part9/requirements.txt
appdirs==1.4.3
certifi==2018.4.16
chardet==3.0.4
coverage==3.7.1
decorator==4.3.0
deprecation==2.0.2
dogpile.cache==0.6.5
ecdsa==0.13
Fabric==1.10.1
Flask==0.10.1
Flask-Bcrypt==0.6.2
Flask-SQLAlchemy==2.0
Flask-WTF==0.11
gunicorn==19.3.0
idna==2.6
iso8601==0.1.12
itsdangerous==0.24
Jinja2==2.7.3
jmespath==0.9.3
jsonpatch==1.23
jsonpointer==2.0
keystoneauth1==3.7.0
MarkupSafe==0.23
munch==2.3.2
netifaces==0.10.7
nose==1.3.4
openstacksdk==0.13.0
os-service-types==1.2.0
packaging==17.1
paramiko==1.15.2
pbr==4.0.3
pycrypto==2.6.1
pyparsing==2.2.0
python-bcrypt==0.3.1
PyYAML==3.12
requests==2.18.4
requestsexceptions==1.4.0
six==1.11.0
SQLAlchemy==0.9.9
stevedore==1.28.0
urllib3==1.22
Werkzeug==0.10.1
WTForms==2.0.2
